refactor: Extract FolderComponent
